+++ test/e2e/e2e.go
@@ -47,7 +47,7 @@ import (
 
 const (
 	operatorNamespace = "mysql-operator"
-	releaseName       = "operator"
+	operatorPodName   = "mysql-operator-0"
 
 	orchestratorPort = 3000
 )
@@ -76,12 +76,12 @@ var _ = ginkgo.SynchronizedBeforeSuite(func() []byte {
 			ginkgo.Fail(fmt.Sprintf("can't create mysql-operator namespace: %s", err))
 		}
 	}
-	framework.HelmInstallChart(releaseName, operatorNamespace)
+	framework.ApplyOperatorManifests(operatorNamespace)
 
 	// Create a tunnel, port-forward orchestrator port to local port
 	ginkgo.By("Port-forward orchestrator")
 	orcTunnel = pf.NewTunnel(restClient, kubeCfg, operatorNamespace,
-		fmt.Sprintf("%s-mysql-operator-0", releaseName),
+		operatorPodName,
 		orchestratorPort,
 	)
 	if err := orcTunnel.ForwardPort(); err != nil {
@@ -119,8 +119,8 @@ var _ = ginkgo.SynchronizedAfterSuite(func() {
 	client, err := clientset.NewForConfig(kubeCfg)
 	gomega.Expect(err).NotTo(gomega.HaveOccurred())
 
-	ginkgo.By("Remove operator release")
-	framework.HelmPurgeRelease(releaseName, operatorNamespace)
+	ginkgo.By("Remove operator manifests")
+	framework.RemoveOperatorManifests(operatorNamespace)
 
 	ginkgo.By("Delete operator namespace")